2005/06/15 12:55:22+02:00 utwente.nl!koen
sapwood_1.20-1.bb:  extra install_append, courtesy Florian Boor
osso-gnome-vfs2_2.8.4.4-1.bb:  adjust depends, courtesy Florian Boor
Many files:  added, courtesy Florian Boor
hildon-libs_0.9.6-2.bb:  adjust staging, courtesy Florian Boor

2005/06/15 12:18:24+02:00 uni-frankfurt.de!mickeyl
add mathomatic, a small, portable, general purpose CAS. Patch courtesy Simon Pickering

2005/06/15 11:58:18+02:00 utwente.nl!koen
Many meamo files:  source -> ., courtesy Florian Boor
Add gconf-osso, part of the maemo.org suite, courtesy Florian Boor

2005/06/15 11:45:40+02:00 utwente.nl!koen
base-files_3.0.14.bb:  support nokia770, courtesy Florian Boor

2005/06/15 11:05:24+02:00 handhelds.org!CoreDump
Merge bk://oe-devel@oe-devel.bkbits.net/openembedded
into handhelds.org:/home/mhentges/OpenEmbedded/bitbake/openembedded

2005/06/15 11:05:02+02:00 handhelds.org!CoreDump
altboot: Add a timeout on boot and only show the menu when requested

2005/06/15 10:29:53+02:00 mn-solutions.de!schurig
assorted MNCI / RT3000 changes

2005/06/15 10:27:59+02:00 mn-solutions.de!schurig
classes/rootfs_ipk.bbclass:
allows to keep the Packages file
(my own python script to create Packages is way faster :-)

2005/06/14 22:59:08+01:00 reciva.com!pb
use mkdir -p, not install -d

2005/06/14 22:02:28+02:00 utwente.nl!koen
Merge bk://oe-devel@oe-devel.bkbits.net/openembedded
into bitbake.utwente.nl:/home/koen/OE/openembedded

2005/06/14 22:02:15+02:00 utwente.nl!koen
xmame_0.95.bb:
  remove UNNECESSARY and BOGUS qpe crap. Can we please stop committing such nonsonse? virtual/libsdl exists for a reason

2005/06/14 21:01:19+01:00 reciva.com!pb
apply PARALLEL_INSTALL_MODULES patch from Erik Hovland


2005/06/14 20:55:08+01:00 reciva.com!pb
update xmame to 0.95, courtesy Erik Hovland


2005/06/14 20:52:19+01:00 reciva.com!pb
update handhelds-pxa kernel to hh40, courtesy Erik Hovland


2005/06/14 19:10:13+02:00 uni-frankfurt.de!mickeyl
Merge bk://oe-devel@oe-devel.bkbits.net/openembedded
into r2d2.tm.informatik.uni-frankfurt.de:/local/pkg/oe/packages

2005/06/14 19:00:35+02:00 uni-frankfurt.de!mickeyl
Add tslib config files for omap5912osk. Patch courtesy philip@balister.org

2005/06/14 18:59:02+02:00 uni-frankfurt.de!mickeyl
Fix Typo in xerces-c. Patch courtesy philip@balister.org

2005/06/14 13:45:03+02:00 local!hrw
Merge bk://oe-devel@oe-devel.bkbits.net/openembedded
into marcinj.local:/home/hrw/zaurus/oe/openembedded

2005/06/14 13:44:34+02:00 local!hrw
updated kdepimpi to 2.1.5

2005/06/14 13:16:38+02:00 handhelds.org!CoreDump
Merge bk://oe-devel@oe-devel.bkbits.net/openembedded
into handhelds.org:/home/mhentges/OpenEmbedded/bitbake/openembedded

2005/06/14 13:16:14+02:00 handhelds.org!CoreDump
altboot: Update to latest version

2005/06/14 13:13:37+02:00 utwente.nl!koen
Merge bk://oe-devel@oe-devel.bkbits.net/openembedded
into bitbake.utwente.nl:/home/koen/OE/openembedded

2005/06/14 13:13:23+02:00 utwente.nl!koen
Add tune-supersparc.conf for people who have v8 cpu's and want hardware integer multiply and divide. Makes ssh login approx. 30 times faster

2005/06/14 11:55:00+01:00 reciva.com!pb
Merge bk://oe-devel@openembedded-devel.bkbits.net/openembedded
into mill.internal.reciva.com:/nexus/home/pb/oe/oe

2005/06/14 11:52:01+01:00 reciva.com!pb
eliminate spare copy of sun4cdm kernel preference


2005/06/14 12:43:33+02:00 mn-solutions.de!schurig
Merge bk://oe-devel@oe-devel.bkbits.net/openembedded/
into mnz66.mn-solutions.de:/usr/src/mnci54/oe

2005/06/14 12:43:12+02:00 mn-solutions.de!schurig
renamed MACHINE 'ramses' to 'mnci'
added new MACHINE 'rt3000'

renamed DISTRO 'openmnci' to 'openmn', which now is for both MNCI and RT3000

2005/06/14 09:52:19+02:00 mn-solutions.de!schurig
Merge bk://oe-devel@oe-devel.bkbits.net/openembedded/
into mnz66.mn-solutions.de:/usr/src/mnci54/oe

2005/06/10 09:37:26+02:00 mn-solutions.de!schurig
Merge bk://oe-devel@oe-devel.bkbits.net/openembedded/
into mnz66.mn-solutions.de:/usr/src/mnci54/oe

2005/06/08 10:05:02+02:00 mn-solutions.de!schurig
Merge bk://oe-devel@oe-devel.bkbits.net/openembedded/
into mnz66.mn-solutions.de:/usr/src/mnci54/oe

2005/06/08 09:16:06+02:00 mn-solutions.de!schurig
icewm: use sourceforge mirror

BKrev: 42b037e5CrddHepCs3gW5YAsfOQ7ng

+++ classes/rootfs_ipk.bbclass
@@ -30,9 +30,11 @@ real_do_rootfs () {
 		
 	mkdir -p ${IMAGE_ROOTFS}/dev
 
-	rm -f ${DEPLOY_DIR_IPK}/Packages
-	touch ${DEPLOY_DIR_IPK}/Packages
-	ipkg-make-index -r ${DEPLOY_DIR_IPK}/Packages -p ${DEPLOY_DIR_IPK}/Packages -l ${DEPLOY_DIR_IPK}/Packages.filelist -m ${DEPLOY_DIR_IPK}
+	if -z "${DEPLOY_KEEP_PACKAGES}"; then
+		rm -f ${DEPLOY_DIR_IPK}/Packages
+		touch ${DEPLOY_DIR_IPK}/Packages
+		ipkg-make-index -r ${DEPLOY_DIR_IPK}/Packages -p ${DEPLOY_DIR_IPK}/Packages -l ${DEPLOY_DIR_IPK}/Packages.filelist -m ${DEPLOY_DIR_IPK}
+	fi
 	mkdir -p ${T}
 	echo "src oe file:${DEPLOY_DIR_IPK}" > ${T}/ipkg.conf
 	ipkgarchs="all any noarch ${TARGET_ARCH} ${IPKG_ARCHS} ${MACHINE}"
